Move the definition of the URI type to the endpoints module, as it is coupled to the endponts (specifically tied to an endpoint, and checks for an endpoint path in its validation logic)
