+++ src/blueprints/dataset.py
@@ -10,7 +10,7 @@ from pydantic import BaseModel, Field
 
 from src.blueprints.common_models import GroupACL, ParameterSet, UserACL
 from src.blueprints.custom_data_types import ISODateTime, MTUrl
-from src.mytardis_client.data_types import URI
+from src.mytardis_client.endpoints import URI
 from src.mytardis_client.enumerators import DataClassification, DataStatus
