+++ src/mytardis_client/mt_rest.py
@@ -16,7 +16,8 @@ from requests import Response
 from requests.exceptions import RequestException
 
 from src.config.config import AuthConfig, ConnectionConfig
-from src.mytardis_client.data_types import URI, HttpRequestMethod
+from src.mytardis_client.endpoints import URI
+, HttpRequestMethod
 from src.mytardis_client.endpoints import MyTardisEndpoint
 
 # Defines the valid values for the MyTardis API version
