+++ src/overseers/overseer.py
@@ -9,8 +9,7 @@ from typing import Any
 from pydantic import ValidationError
 from requests.exceptions import HTTPError
 
-from src.mytardis_client.data_types import URI
-from src.mytardis_client.endpoints import MyTardisEndpoint
+from src.mytardis_client.endpoints import URI, MyTardisEndpoint
 from src.mytardis_client.mt_rest import MyTardisRESTFactory
 from src.mytardis_client.objects import MyTardisObject, get_type_info
 from src.mytardis_client.response_data import MyTardisIntrospection
