+++ tests/fixtures/fixtures_dataclasses.py
@@ -19,7 +19,7 @@ from src.blueprints.datafile import (
 from src.blueprints.dataset import Dataset, RawDataset, RefinedDataset
 from src.blueprints.experiment import Experiment, RawExperiment, RefinedExperiment
 from src.blueprints.project import Project, RawProject, RefinedProject
-from src.mytardis_client.data_types import URI
+from src.mytardis_client.endpoints import URI
 from src.mytardis_client.enumerators import DataClassification
