+++ tests/test_dataclasses.py
@@ -6,7 +6,7 @@ from pytest import fixture
 
 from src.blueprints.common_models import Parameter, ParameterSet
 from src.blueprints.custom_data_types import MTUrl
-from src.mytardis_client.data_types import URI
+from src.mytardis_client.endpoints import URI
 
 
 @fixture
