+++ tests/test_mytardis_client_data_types.py
@@ -2,7 +2,7 @@
 # nosec assert_used
 import pytest
 
-from src.mytardis_client.data_types import URI, resource_uri_to_id
+from src.mytardis_client.endpoints import URI, resource_uri_to_id
 from src.mytardis_client.objects import KNOWN_MYTARDIS_OBJECTS
